articlegenerator/requirements.txt
# Core Django requirements
Django==4.2  # Replace with your Django version
djangorestframework==3.14.0  # For Django REST Framework

# Celery and related dependencies
celery[redis]==5.3.1  # For Celery and Redis integration
django-celery-results==2.5.1  # For storing Celery task results in the database

# Web scraping dependencies
requests==2.31.0  # For HTTP requests
beautifulsoup4==4.12.2  # For HTML parsing

# Feed parsing
feedparser==6.0.10  # For parsing RSS/Atom feeds

# Machine learning dependencies
scikit-learn  # For KMeans and silhouette score

# PostgreSQL database
psycopg2-binary==2.9.8  # For PostgreSQL integration

# pgvector integration
pgvector  # For vector operations in PostgreSQL

# Additional optional dependencies
gunicorn==20.1.0  # If deploying with Gunicorn

numpy

django-extensions

django-celery-beat